+++ OUDS/Core/Components/Sources/Indicators/ProgressIndicator/Internal/CircularProgressIndicator/CircularProgressIndicatorView.swift
@@ -56,7 +56,6 @@ struct CircularProgressIndicatorView: View {
                         trackColor: trackColor,
                         strokeCap: strokeCap,
                         gapSize: configuration.gapSize,
-                        hasTrack: configuration.track,
                         size: scaledDefaultSize)
                 }
             }