+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Animations for `progress indicator` components (Orange-OpenSource/ouds-ios#1708)
 - Update icons to v2.3.0 (Orange-OpenSource/ouds-ios#1432)
 - `progress indicator` components to version 1.2.0 (Orange-OpenSource/ouds-ios#1674)
 - `password input` component to version 1.3.1 (Orange-OpenSource/ouds-ios#1526)